Restructure and refactoring
