+++ src/main/java/net/explorviz/adapter/conversion/converter/SpanDynamicConverter.java
@@ -1,4 +1,4 @@
-package net.explorviz.adapter.translation;
+package net.explorviz.adapter.conversion.converter;
 
 import io.opencensus.proto.trace.v1.Span;
 import javax.enterprise.context.ApplicationScoped;
