+++ src/main/java/net/explorviz/adapter/conversion/transformer/StructureTransformer.java
@@ -1,12 +1,11 @@
-package net.explorviz.adapter.kafka;
+package net.explorviz.adapter.conversion.transformer;
 
 import io.opencensus.proto.trace.v1.Span;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
-import net.explorviz.adapter.translation.SpanStructureConverter;
+import net.explorviz.adapter.conversion.converter.SpanStructureConverter;
 import net.explorviz.adapter.util.PerfomanceLogger;
 import net.explorviz.adapter.validation.SpanStructureSanitizer;
-import net.explorviz.adapter.validation.SpanValidator;
 import net.explorviz.avro.SpanStructure;
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.kstream.Transformer;
